fix(qty): P1 — request quantity Decimal (Law 25) + ItemPreview chat + clarify jaga desimal
Unit P1 (hygiene Law 25 + pemotong DISPLAY chat), atas putusan MASTER.
Gerbang dua sisi 8/8 GREEN.

TEMUAN: jalur TULIS chat TIDAK memotong qty (terverifikasi) — chat "0,51 kg" sudah
menulis 0.51 ke DB dgn benar. Pemotong yg tersisa hanya di DISPLAY.

- schemas request quantity float -> Decimal (Law 25 "Decimal di model, float di respons"):
  purchase_orders (Create+Update), quotes, sales_orders, credit_notes, vendor_credits,
  stock_adjustments (quantity_adjustment + physical_quantity). Konsumen AMAN: nol float-arith
  pada .quantity di router (audit). float lama SUDAH terima 0.51 → ini presisi/konsistensi,
  BUKAN perbaikan truncation; nol regresi. Respons quantity tetap float (Law 25).
- schemas/action_chat.py ItemPreview qty/price/subtotal int -> float. Pydantic v2 int MENOLAK
  0.51 (int_from_float); kartu preview chat kini bisa tampil 0.51.
- unified_agent/tool_executor.py:4210 clarify qty int(float(...)) -> jaga desimal (idiom sama
  dgn :5825 int(_q) if _q==int(_q) else _q). Ini kartu KLARIFIKASI (tanya qty), bukan tulis dok.

Sisa display minor (backlog, tak di commit ini): action_memory.py:251 int(qty) "%d pcs".

+++ backend/api_gateway/app/schemas/action_chat.py
@@ -49,11 +49,11 @@ class CancelActionRequest(BaseModel):
 
 class ItemPreview(BaseModel):
     name: str
-    qty: int
+    qty: float
     unit: str
-    price: int
+    price: float
     discount_percent: float = 0
-    subtotal: int
+    subtotal: float
 
 
 class CalculationPreview(BaseModel):
